Moved Close account functionality into a new separate controller
